feat: implement Recording Overlay (TASK-2.2)
- RecordingOverlayWindow: frameless, transparent, topmost, not in taskbar
- ShowActivated=False to prevent focus stealing during recording
- Draggable via BeginMoveDrag; position saved to Settings (OverlayX/OverlayY)
- Default placement at bottom-center of screen on first launch
- RecordingOverlayViewModel: binds to IAudioRecorder state/duration
- MM:SS duration timer updated via DispatcherTimer
- Pulsing red dot + animated waveform bars (CSS animations)
- StopRecordingCommand delegates to IAudioRecorder.StopRecording()
- IDisposable implemented for clean event handler unsubscription
- EF migration AddOverlayPosition adds OverlayX/OverlayY (int?) to Settings
- 19 new unit tests (424 total, all passing)

+++ source/VivaVoz/Migrations/20260223150000_AddOverlayPosition.cs
@@ -10,16 +10,16 @@ namespace VivaVoz.Migrations;
 public partial class AddOverlayPosition : Migration {
     /// <inheritdoc />
     protected override void Up(MigrationBuilder migrationBuilder) {
-        migrationBuilder.AddColumn<double>(
+        migrationBuilder.AddColumn<int>(
             name: "OverlayX",
             table: "Settings",
-            type: "REAL",
+            type: "INTEGER",
             nullable: true);
 
-        migrationBuilder.AddColumn<double>(
+        migrationBuilder.AddColumn<int>(
             name: "OverlayY",
             table: "Settings",
-            type: "REAL",
+            type: "INTEGER",
             nullable: true);
     }
