+++ source/VivaVoz/Models/Settings.cs
@@ -14,8 +14,8 @@ public class Settings {
     public bool StartMinimized { get; set; } = false;
     public bool RunAtStartup { get; set; } = false;
     public string RecordingMode { get; set; } = "Toggle";
-    public double? OverlayX { get; set; }
-    public double? OverlayY { get; set; }
+    public int? OverlayX { get; set; }
+    public int? OverlayY { get; set; }
 
     private static string GetDefaultStoragePath() {
         var localAppData = Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData);
